organised code into more folders and files
modularity has been mostly achieved.

+++ src/types/types.ts
@@ -2,5 +2,5 @@ export type MeshData = {
   positions: Float32Array | number[];
   uvs: Float32Array | number[];
   normals: Float32Array | number[];
-  indices: Uint16Array | number[];
+  indices: Uint32Array | number[];
 };
